First step to unify Koda's binary op type promotion logic and the DType lattice.
Made lattice to be a matrix to make it simple to use constexpr.

I also had to use std::array<bool> instead of std::bitset due to better constexpr support.

TODO: we may also consider to reuse common type in case we want to be able to find common upper bound.
PiperOrigin-RevId: 810019897
Change-Id: I5635a95134fe31e2f87b68c8288c1dec89fb8442