[RFC] Prevent recursive functor calls from segfaulting
This causes the python kernel to crash, and is generally a bad thing to happen.

Partially inspired by protobuf, which limits proto parsing depth to 100. Given the large discrepancy in stack size between e.g. kd.streams.call and kd.call, I've tried to be a bit more dynamic. Note that `pthread_attr_getstacksize` is not available on e.g. Windows, but AFAIU we don't support it anyways.

PiperOrigin-RevId: 812676991
Change-Id: I3dd600eaa96bfab1d947eb936af668fff72c334d