Add repr to View
Helps with debugging and quick inspection without having to do `.map(print)`.

We may also consider whether we should instead print something like `view(((1, 2), (3,))[:][:]` which is closer to what the user may have written. However, this is likely _not_ exactly what they have done, as there are many ways to obtain the same state so I think this is likely more confusing than not. When / if we have paths, this could potentially be integrated instead.

PiperOrigin-RevId: 823452863
Change-Id: I0d2783cfac4ef4b81d09459fca1d2c163ced4cb8