Share test data for kv.map operator with Koda.
Add support for ndim= in kv.map, and remove agg_map since ndim= is a more general version of that.

It is a corner case that kv.map corresponds to kd.map_py and not to kd.map, but it would seem weird to have "_py" suffix in Koda View where everything is a Python object anyway.

PiperOrigin-RevId: 825050465
Change-Id: Ia527f4732fc06184d1d090859c1a9b11ff128a8c