Change set_attr to set_attrs
Rather than having the callback take `**attrs`, we repeatedly call map2 with a single attr instead. This is much faster when you just have a few attrs that need to be updated, since we avoid repeatedly looping over dict keys / values.

It also allows us to use it in `__setattr__` with paying much of a cost and without having a custom impl.

PiperOrigin-RevId: 827978918
Change-Id: Idd8afff22795071b54417482e85a7b97bc104245