Add TransformTo[Simple]ParallelCallFn()
The goal of this change is to provide a simplified API for using parallel transformed functors in C++, making them easier to use without knowing the internals.

`TransformToSimpleParallelCallFn(functor_data_slice)` returns a C++ functor that takes non-parallel values and returns a stream.

`TransformToParallelCallFn(functor_data_slice) is a lower-level primitive that also constructs a C++ functor; however, it leaves the parallel transform of the arguments to the user.

PiperOrigin-RevId: 831022420
Change-Id: I765e37494cc6d1e7e1e7ab4f1ba160b6190c00a0