Refactor Koladata pseudo-random generator to use a weak symbol for reproducibility
This patch replaces the mechanism used to de-randomize Koladata. Instead of relying on an environment variable, it now uses a weak symbol `_KoladataInternalPseudoRandomUint64`.

This allows the behaviour to be controlled by adding a build dependency or by using LD_PRELOAD. The primary motivation is to remove "special" mode logic from client applications. Specifically, the overloaded version can enforce single-thread access to the random number generator without imposing this logic on the default implementation.

PiperOrigin-RevId: 843637091
Change-Id: Ia2b8fb2348b0529138dadb8ce1f3d7cb41138c42