[RFC] Add size and present_count to DataSlice repr.
This improves debuggability and overview of the slice - especially in the case of large slices. This is useful to e.g. see how many values were filtered using select or `&`. The size I find especially useful, so I would also be OK moving the present_count to the debug repr instead if you think this is too much for the regular repr.

PiperOrigin-RevId: 845824818
Change-Id: I979e42e5b1514853345b34db2996b507d61d9f3a